Some changes in structure
